backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
mysqlclient==2.2.1
dj-database-url==2.1.0

# Security & CORS
django-cors-headers==4.3.1
python-decouple==3.8

# Password hashing
argon2-cffi==23.1.0

# Database connections
SQLAlchemy==2.0.25
pandas==2.2.0

# Development
python-dotenv==1.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0
